+++ cmd/rabtap/command_line.go
@@ -21,21 +21,21 @@ const (
 
 Usage:
   rabtap -h|--help
-  rabtap tap EXCHANGES [--uri URI] [--saveto DIR] [-jknv]
-  rabtap (tap --uri URI EXCHANGES)... [--saveto DIR] [-jknv]
-  rabtap info [--api APIURI] [--consumers] [--stats]
-              [--filter EXPR] [--omit-empty] [--show-default]
-              [--mode MODE] [--format FORMAT] [-knv]
-  rabtap pub [--uri URI] EXCHANGE [FILE] [--routingkey=KEY] [-jkv]
-  rabtap sub QUEUE [--uri URI] [--saveto DIR] [--no-auto-ack] [-jkvn]
-  rabtap exchange create EXCHANGE [--uri URI] [--type TYPE] [-adkv]
-  rabtap exchange rm EXCHANGE [--uri URI] [-kv]
-  rabtap queue create QUEUE [--uri URI] [-adkv]
-  rabtap queue bind QUEUE to EXCHANGE --bindingkey=KEY [--uri URI] [-kv]
-  rabtap queue unbind QUEUE from EXCHANGE --bindingkey=KEY [--uri URI] [-kv]
-  rabtap queue rm QUEUE [--uri URI] [-kv]
-  rabtap queue purge QUEUE [--uri URI] [-kv]
-  rabtap conn close CONNECTION [--reason=REASON] [--api APIURI] [-kv]
+  rabtap info [--api=APIURI] [--consumers] [--stats]
+              [--filter=EXPR] [--omit-empty] [--show-default]
+              [--mode=MODE] [--format=FORMAT] [-knv]
+  rabtap tap EXCHANGES [--uri=URI] [--saveto=DIR] [--format=FORMAT] [-jknv]
+  rabtap (tap --uri=URI EXCHANGES)... [--saveto=DIR] [--format=FORMAT] [-jknv]
+  rabtap pub [--uri=URI] EXCHANGE [FILE] [--routingkey=KEY] [--format=FORMAT] [-jkv]
+  rabtap sub QUEUE [--uri URI] [--saveto=DIR] [--format=FORMAT] [--no-auto-ack] [-jkvn]
+  rabtap exchange create EXCHANGE [--uri=URI] [--type=TYPE] [-adkv]
+  rabtap exchange rm EXCHANGE [--uri=URI] [-kv]
+  rabtap queue create QUEUE [--uri=URI] [-adkv]
+  rabtap queue bind QUEUE to EXCHANGE --bindingkey=KEY [--uri=URI] [-kv]
+  rabtap queue unbind QUEUE from EXCHANGE --bindingkey=KEY [--uri=URI] [-kv]
+  rabtap queue rm QUEUE [--uri=URI] [-kv]
+  rabtap queue purge QUEUE [--uri=URI] [-kv]
+  rabtap conn close CONNECTION [--api=APIURI] [--reason=REASON] [-kv]
   rabtap --version
 
 Options:
@@ -47,18 +47,21 @@ Options:
  QUEUE                name of a queue.
  CONNECTION           name of a connection.
  -a, --autodelete     create auto delete exchange/queue.
- --api APIURI         connect to given API server. If APIURI is omitted,
+ --api=APIURI         connect to given API server. If APIURI is omitted,
                       the environment variable RABTAP_APIURI will be used.
- -b, --bindingkey KEY binding key to use in bind queue command.
+ -b, --bindingkey=KEY binding key to use in bind queue command.
  --by-connection      output of info command starts with connections.
  --consumers          include consumers and connections in output of info command.
  -d, --durable        create durable exchange/queue.
- --filter EXPR        Predicate for info command to filter queues [default: true]
+ --filter=EXPR        Predicate for info command to filter queues [default: true]
                       (see README.md for details)
+ --format=FORMAT      * for tap, pub, sub command: format to write/read messages to console
+                      and optionally to file (when --saveto DIR is given). 
+					  Valid options are: "raw", "json", "json-nopp". Default: raw
+					  * for info command: controls generated output format. Valid 
+					  options are: "text", "dot". Default: text
  -h, --help           print this help.
- -j, --json           print/save/publish message metadata and body to a
-                      single JSON file. JSON body is base64 encoded. Otherwise
-                      metadata and body (as-is) are saved separately.
+ -j, --json           Deprecated. Use "--format json" instead.
  -k, --insecure       allow insecure TLS connections (no certificate check).
  --mode=MODE          mode for info command. One of "byConnection", "byExchange".
                       [default: byExchange].
@@ -67,16 +70,14 @@ Options:
                       unacked messages on the broker which will be requeued
                       when the channel is closed.
  -o, --omit-empty     don't show echanges without bindings in info command.
- --format=FORMAT      output format for info command. One of "text", "dot".
-                      [default: text]
  --reason=REASON      reason why the connection was closed
                       [default: closed by rabtap].
- -r, --routingkey KEY routing key to use in publish mode.
- --saveto DIR         also save messages and metadata to DIR.
+ -r, --routingkey=KEY routing key to use in publish mode.
+ --saveto=DIR         also save messages and metadata to DIR.
  --show-default       include default exchange in output info command.
  --stats              include statistics in output of info command.
- -t, --type TYPE      exchange type [default: fanout].
- --uri URI            connect to given AQMP broker. If omitted, the
+ -t, --type=TYPE      exchange type [default: fanout].
+ --uri=URI            connect to given AQMP broker. If omitted, the
                       environment variable RABTAP_AMQPURI will be used.
  -v, --verbose        enable verbose mode.
  --version            show version information and exit.
@@ -137,7 +138,6 @@ type commonArgs struct {
 	Verbose     bool
 	InsecureTLS bool
 	NoColor     bool
-	Format      string // output format subscribe, tap: "raw", "json"
 	AmqpURI     string // pub, queue, exchange: amqp broker to use
 }
 
@@ -164,7 +164,7 @@ type CommandLineArgs struct {
 	QueueFilter         string  // info mode: optional filter predicate
 	OmitEmptyExchanges  bool    // info mode: do not show exchanges wo/ bindings
 	ShowDefaultExchange bool    // info mode: show default exchange
-	InfoFormat          string  // info mode: output format
+	Format              string  // output format, depends on command
 	Durable             bool    // queue create, exchange create
 	Autodelete          bool    // queue create, exchange create
 	SaveDir             *string // save mode: optional directory to stores files to
@@ -210,15 +210,10 @@ func parseAPIURI(args map[string]interface{}) (string, error) {
 }
 
 func parseCommonArgs(args map[string]interface{}) commonArgs {
-	format := "raw"
-	if args["--json"].(bool) {
-		format = "json"
-	}
 	return commonArgs{
 		Verbose:     args["--verbose"].(bool),
 		InsecureTLS: args["--insecure"].(bool),
-		NoColor:     args["--no-color"].(bool) || (os.Getenv("NO_COLOR") != ""),
-		Format:      format}
+		NoColor:     args["--no-color"].(bool) || (os.Getenv("NO_COLOR") != "")}
 }
 
 func parseInfoCmdArgs(args map[string]interface{}) (CommandLineArgs, error) {
@@ -237,11 +232,14 @@ func parseInfoCmdArgs(args map[string]interface{}) (CommandLineArgs, error) {
 	}
 	result.InfoMode = mode
 
-	format := args["--format"].(string)
+	format := "text"
+	if args["--format"] != nil {
+		format = args["--format"].(string)
+	}
 	if format != "text" && format != "dot" {
 		return result, errors.New("--format FORMAT must be one of {text, dot}")
 	}
-	result.InfoFormat = format
+	result.Format = format
 
 	var err error
 	if result.APIURI, err = parseAPIURI(args); err != nil {
@@ -266,6 +264,25 @@ func parseConnCmdArgs(args map[string]interface{}) (CommandLineArgs, error) {
 	return result, nil
 }
 
+// parsePubSubFormatArg parse --format option for pub, sub, tap command.
+func parsePubSubFormatArg(args map[string]interface{}) (string, error) {
+	format := "raw"
+
+	if args["--format"] != nil {
+		format = args["--format"].(string)
+	}
+
+	// deprecated --json option equals "--format json"
+	if args["--json"] != nil && args["--json"].(bool) {
+		format = "json"
+	}
+
+	if format != "json" && format != "json-nopp" && format != "raw" {
+		return "", errors.New("--format FORMAT must be one of {raw,json,json-nopp}")
+	}
+	return format, nil
+}
+
 func parseSubCmdArgs(args map[string]interface{}) (CommandLineArgs, error) {
 	result := CommandLineArgs{
 		Cmd:        SubCmd,
@@ -273,7 +290,13 @@ func parseSubCmdArgs(args map[string]interface{}) (CommandLineArgs, error) {
 		AutoAck:    !args["--no-auto-ack"].(bool),
 		QueueName:  args["QUEUE"].(string),
 	}
-	var err error
+
+	format, err := parsePubSubFormatArg(args)
+	if err != nil {
+		return result, err
+	}
+	result.Format = format
+
 	if args["--saveto"] != nil {
 		saveDir := args["--saveto"].(string)
 		result.SaveDir = &saveDir
@@ -342,7 +365,12 @@ func parsePublishCmdArgs(args map[string]interface{}) (CommandLineArgs, error) {
 		Cmd:        PubCmd,
 		commonArgs: parseCommonArgs(args)}
 
-	var err error
+	format, err := parsePubSubFormatArg(args)
+	if err != nil {
+		return result, err
+	}
+	result.Format = format
+
 	if result.AmqpURI, err = parseAmqpURI(args); err != nil {
 		return result, err
 	}
@@ -362,6 +390,12 @@ func parseTapCmdArgs(args map[string]interface{}) (CommandLineArgs, error) {
 		Cmd:        TapCmd,
 		commonArgs: parseCommonArgs(args),
 		TapConfig:  []rabtap.TapConfiguration{}}
+
+	format, err := parsePubSubFormatArg(args)
+	if err != nil {
+		return result, err
+	}
+	result.Format = format
 
 	if args["--saveto"] != nil {
 		saveDir := args["--saveto"].(string)